+++ imap_processing/ultra/l0/ultra_utils.py
@@ -1,39 +1,39 @@
-"""Data class for Ultra Level 0 data."""
-from enum import IntEnum
+"""Contains data classes to support Ultra L0 processing."""
+from enum import Enum
+from typing import NamedTuple
 from imap_processing.cdf.defaults import GlobalConstants
+from dataclasses import fields
 
 
-class ULTRAAPID(IntEnum):
-    """Create ENUM for apid.
+class PacketProperties(NamedTuple):
+    """Class that represents properties of the ULTRA packet type."""
 
-    Parameters
-    ----------
-    IntEnum : IntEnum
-    """
-
-    # Single Events: 0x380, 0x3c0
-    ULTRA_EVENTS_45 = 896
-    ULTRA_EVENTS_90 = 960
-
-    # Auxiliary Data: 0x370, 0x3b0
-    ULTRA_AUX_45 = 880
-    ULTRA_AUX_90 = 994
-
-    # Image Rates: 0x371, 0x3b1
-    ULTRA_RATES_45 = 881
-    ULTRA_RATES_90 = 945
-
-    # TOF Images: 0x373,0x3b3
-    ULTRA_TOF_45 = 883
-    ULTRA_TOF_90 = 947
+    apid: int
+    width: int
+    block: int
+    len_array: int
+    mantissa_bit_length: int
 
 
-class EventParser:
+class UltraParams(Enum):
+    """Enumerated packet properties for ULTRA."""
+
+    ULTRA_AUX = PacketProperties(
+        apid=[880,994], width=None, block=None, len_array=None, mantissa_bit_length=None)
+    ULTRA_RATES = PacketProperties(
+        apid=[881,945], width=5, block=16, len_array=48, mantissa_bit_length=12)
+    ULTRA_TOF = PacketProperties(
+        apid=[883,947], width=4, block=15, len_array=None, mantissa_bit_length=4)
+    ULTRA_EVENTS = PacketProperties(
+        apid=[896,960], width=None, block=None, len_array=None, mantissa_bit_length=None)
+
+
+class ParserHelper:
     """Data class for parsing Ultra Level 0 event data from binary packets.
 
     Attributes
     ----------
-    field_ranges : dict
+    event_field_ranges : dict
         A dictionary mapping event field names to their respective start and end
         bit positions within a binary event string.
     scalar_fields : set
@@ -57,7 +57,7 @@ class EventParser:
         given packet, updating the event data structure.
     """
     def __init__(self):
-        self.field_ranges = {
+        self.event_field_ranges = {
             "coin_type": (0, 2),
             "start_type": (2, 4),
             "stop_type": (4, 8),
@@ -77,12 +77,12 @@ class EventParser:
             "bin": (148, 156),
             "phase_angle": (156, 166),
         }
-        self.event_scalar_fields = {'SHCOARSE', 'SID', 'SPIN', 'ABORTFLAG', 'STARTDELAY', 'COUNT'}
+        self.scalar_fields = {'SHCOARSE', 'SID', 'SPIN', 'ABORTFLAG', 'STARTDELAY', 'COUNT'}
 
-    def initialize_event_data(self, header):
+    def initialize_event_data(self):
         """Initializes and returns the data structure for storing event data."""
-        event_data = {field: [] for field in self.field_ranges}
-        for scalar_field in self.event_scalar_fields:
+        event_data = {field: [] for field in self.event_field_ranges}
+        for scalar_field in self.scalar_fields:
             event_data[scalar_field] = []
 
         return event_data
@@ -90,17 +90,25 @@ class EventParser:
     def append_fillval(self, event_data):
         """Appends fillvalue to all event fields except for specified scalar and CCSDS fields."""
         for key in event_data:
-            if key not in self.event_scalar_fields:
+            if key not in self.scalar_fields:
                 event_data[key].append(GlobalConstants.INT_FILLVAL)
 
     def parse_event(self, event_binary):
         """Parses a binary string representing a single event."""
         return {
             field: int(event_binary[start:end], 2)
-            for field, (start, end) in self.field_ranges.items()
+            for field, (start, end) in self.event_field_ranges.items()
         }
 
     def append_values(self, event_data, packet):
         """Appends scalar fields to event_data."""
-        for key in self.event_scalar_fields:
+        for key in self.scalar_fields:
             event_data[key].append(packet.data[key].raw_value)
+
+    def append_ccsds_fields(self, event_data, ccsds_data_object):
+        """Appends ccsds fields to event_data."""
+        for field in fields(ccsds_data_object.__class__):
+            ccsds_key = field.name
+            if ccsds_key not in event_data:
+                event_data[ccsds_key] = []
+            event_data[ccsds_key].append(getattr(ccsds_data_object, ccsds_key))
